refactor: 필드에 final 적용, UpdateProductRequest 의 quantity 를 long 으로 변경

+++ shop/shop-domain/src/main/java/shop/woowasap/shop/dto/UpdateProductRequest.java
@@ -2,7 +2,7 @@ package shop.woowasap.shop.dto;
 
 import java.time.LocalDateTime;
 
-public record UpdateProductRequest(String name, String description, String price, int quantity,
+public record UpdateProductRequest(String name, String description, String price, long quantity,
                                   LocalDateTime startTime, LocalDateTime endTime) {
 
 }
